chore(deps): bump version to 0.1.2-alpha and update dependencies
- Bump package version from 0.1.1-alpha to 0.1.2-alpha across package.json, Cargo.toml, and tauri.conf.json
- Update CodeMirror packages: commands (6.10.2), lint (6.9.3), view (6.39.13)
- Update Tauri packages: api (2.10.1), cli (2.10.0), plugin-shell (2.3.5)
- Update i18next (25.8.4), pdfjs-dist (5.4.624), @vitejs/plugin-react (5.1.4)
- Update Rust dependencies: reqwest (0.13), tokio-tungstenite (0.28), rusqlite (0.38), notify (8

+++ src-tauri/src/tree_sitter/symbol.rs
@@ -216,14 +216,14 @@ impl SymbolExtractor {
 
             // Process children with this symbol as parent
             for i in 0..node.child_count() {
-                if let Some(child) = node.child(i) {
+                if let Some(child) = node.child(i as u32) {
                     self.extract_from_node(child, source, language, Some(&symbol_id), symbols);
                 }
             }
         } else {
             // No symbol at this node, process children with same parent
             for i in 0..node.child_count() {
-                if let Some(child) = node.child(i) {
+                if let Some(child) = node.child(i as u32) {
                     self.extract_from_node(child, source, language, parent_id, symbols);
                 }
             }
